backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-groq
pinecone
pypdf
sentence-transformers
pydantic
python-dotenv
